Refactor layers and utils
