+++ superset-frontend/plugins/legacy-preset-chart-deckgl/package.json
@@ -31,7 +31,7 @@
     "d3-array": "^1.2.4",
     "d3-color": "^1.4.1",
     "d3-scale": "^3.0.0",
-    "deck.gl": "8.5.2",
+    "deck.gl": "8.8.27",
     "lodash": "^4.17.15",
     "moment": "^2.20.1",
     "mousetrap": "^1.6.1",
@@ -40,6 +40,11 @@
     "underscore": "^1.12.1",
     "urijs": "^1.19.8",
     "xss": "^1.0.10"
+  },
+  "devDependencies": {
+    "@types/mapbox__geojson-extent": "^1.0.0",
+    "@types/underscore": "^1.11.6",
+    "@types/urijs": "^1.19.19"
   },
   "peerDependencies": {
     "@superset-ui/chart-controls": "*",
@@ -51,9 +56,5 @@
   },
   "publishConfig": {
     "access": "public"
-  },
-  "devDependencies": {
-    "@types/underscore": "^1.11.6",
-    "@types/urijs": "^1.19.19"
   }
 }
