+++ superset-frontend/plugins/legacy-preset-chart-deckgl/src/utils/colors.ts
@@ -19,7 +19,10 @@
 import { rgb } from 'd3-color';
 
 // eslint-disable-next-line import/prefer-default-export
-export function hexToRGB(hex: string | undefined, alpha = 255) {
+export function hexToRGB(
+  hex: string | undefined,
+  alpha = 255,
+): [number, number, number, number] {
   if (!hex) {
     return [0, 0, 0, alpha];
   }
